tests/test_day10.rs
use advent_of_code::day10::part1;
use advent_of_code::day10::part2;

const INPUT: &str = "[({(<(())[]>[[{[]{<()<>>
[(()[<>])]({[<{<<[]>>(
{([(<{}[<>[]}>{[]{[(<()>
(((({<>}<{<{<>}{[]{[]{}
[[<[([]))<([[{}[[()]]]
[{[{({}]{}}([{[{{{}}([]
{<[[]]>}<{[{[{[]{()[[[]
[<(<(<(<{}))><([]([]()
<{([([[(<>()){}]>(<<{{
<{([{{}}[<[[[<>{}]]]>[]]";

#[test]
fn test_part1() {
    let expected = 26397;
    let result = part1(INPUT.to_string());
    assert_eq!(expected, result);
}

#[test]
fn test_part2() {
    let expected: i64 = 288957;
    let result = part2(INPUT.to_string());
    assert_eq!(expected, result);
}

#[test]
fn test_input_part1() {
    assert_eq!(
        411471,
        part1(include_str!("../inputs/day10.txt").to_string())
    );
}

#[test]
fn test_input_part2() {
    assert_eq!(
        3122628974,
        part2(include_str!("../inputs/day10.txt").to_string())
    );
}
